refactor: 모달을 관리하는 스택 기능 리듀서를 stack reducer로 추상화 단계를 낮추고 파일 분리
